Mise en place des Tokens
Un peu de sécuritée avec l'ajout des jetons de validation
pour les nouveau utilisateur

TODO l'envoi par mail du token

J'ai dû désactiver le EmployeeControler
TODO déplacer le tasktouser dans le TaskControler
	Sinon on ne pourras pas assigner les taches aux utilisateurs

+++ src/main/java/com/gdt/controller/TaskController.java
@@ -46,7 +46,7 @@ public class TaskController {
 	 * @throws BadRequestException
 	 */
 	@GetMapping(path = "/{id}")
-	public Task read(@PathVariable Long id) throws BadRequestException {
+	public Task read(@PathVariable Integer id) throws BadRequestException {
 		return this.service.read(id);
 	}
